requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
asgiref==3.11.0
    # via django
bandit==1.7.10
    # via -r requirements.in
black==24.10.0
    # via -r requirements.in
blinker==1.8.2
    # via flask
build==1.2.2.post1
    # via pip-tools
click==8.1.7
    # via
    #   black
    #   flask
    #   pip-tools
django==6.0.1
    # via
    #   django-filter
    #   djangorestframework
django-filter==25.2
    # via -r requirements.in
djangorestframework==3.16.1
    # via -r requirements.in
flask==3.0.3
    # via -r requirements.in
isort==5.13.2
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markdown==3.10.1
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via
    #   jinja2
    #   werkzeug
mdurl==0.1.2
    # via markdown-it-py
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via
    #   black
    #   build
    #   wheel
pathspec==0.12.1
    # via black
pbr==6.1.0
    # via stevedore
#pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via black
postgres==4.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via
    #   postgres
    #   psycopg2-pool
psycopg2-pool==1.2
    # via postgres
pygments==2.18.0
    # via rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pyyaml==6.0.2
    # via bandit
rich==13.9.2
    # via bandit
sqlparse==0.5.5
    # via django
stevedore==5.3.0
    # via bandit
terminator==0.0.0
    # via -r requirements.in
werkzeug==3.0.6
    # via flask
wheel==0.46.2
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
